cambio de base

+++ Models/Categoria.cs
@@ -10,7 +10,7 @@ namespace Proyecto.Models
             Pregunta = new HashSet<Pregunta>();
         }
 
-        public int Catid { get; set; }
+        public decimal Catid { get; set; }
         public string Catnombre { get; set; }
         public string Catdescripcion { get; set; }
