+++ Models/Respuesta.cs
@@ -5,9 +5,9 @@ namespace Proyecto.Models
 {
     public partial class Respuesta
     {
-        public int Respid { get; set; }
-        public int Userid { get; set; }
-        public int Pregid { get; set; }
+        public decimal Respid { get; set; }
+        public decimal Userid { get; set; }
+        public decimal Pregid { get; set; }
         public string Resptexto { get; set; }
         public DateTime Respfecha { get; set; }
         public TimeSpan Resphora { get; set; }
